fix: started refactoring the UserPreferences service and added env variable for language
