+++ src/js/app.services.js
@@ -6,7 +6,6 @@ import "./services/appointments.service.js";
 import "./services/firebase.service.js";
 import "./services/utility.service";
 import "./services/notificationsService.js";
-import "./services/userPreferencesService.js";
 import "./services/userAuthorizationInfoService.js";
 import "./services/updateUIService.js";
 import "./services/documentsService.js";
@@ -45,4 +44,5 @@ import "./services/versionService.js";
 import "./services/hospital.service.js";
 import "./services/profile-selector.service";
 import "./services/user.service";
+import "./services/user-preferences.service.js";
 import "./services/studiesService.js";
